+++ test/integration/test_b2_command_line.py
@@ -16,12 +16,12 @@ import json
 import os
 import os.path
 import re
+import sys
 
 from pathlib import Path
 from typing import Optional, Tuple
 
 import pytest
-
 from b2sdk.v2 import B2_ACCOUNT_INFO_ENV_VAR, SSE_C_KEY_ID_FILE_INFO_KEY_NAME, UNKNOWN_FILE_RETENTION_SETTING, EncryptionMode, EncryptionSetting, FileRetentionSetting, LegalHold, RetentionMode, fix_windows_path_limit
 
 from b2.console_tool import current_time_millis
@@ -89,6 +89,16 @@ def test_basic(b2_tool, bucket_name):
             file_to_upload, 'd'
         ]
     )
+
+    b2_tool.should_succeed(['upload-file', '--noProgress', bucket_name, file_to_upload, 'rm'])
+    b2_tool.should_succeed(['upload-file', '--noProgress', bucket_name, file_to_upload, 'rm1'])
+    # with_wildcard allows us to target a single file. rm will be removed, rm1 will be left alone
+    b2_tool.should_succeed(['rm', '--recursive', '--withWildcard', bucket_name, 'rm'])
+    list_of_files = b2_tool.should_succeed_json(
+        ['ls', '--json', '--recursive', '--withWildcard', bucket_name, 'rm*']
+    )
+    should_equal(['rm1'], [f['fileName'] for f in list_of_files])
+    b2_tool.should_succeed(['rm', '--recursive', '--withWildcard', bucket_name, 'rm1'])
 
     with TempDir() as dir_path:
         b2_tool.should_succeed(
@@ -231,6 +241,8 @@ def test_key_restrictions(b2_api, b2_tool, bucket_name):
 
     second_bucket_name = b2_tool.generate_bucket_name()
     b2_tool.should_succeed(['create-bucket', second_bucket_name, 'allPublic', *get_bucketinfo()],)
+    # A single file for rm to fail on.
+    b2_tool.should_succeed(['upload-file', '--noProgress', bucket_name, 'README.md', 'test'])
 
     key_one_name = 'clt-testKey-01' + random_hex(6)
     created_key_stdout = b2_tool.should_succeed(
@@ -267,11 +279,22 @@ def test_key_restrictions(b2_api, b2_tool, bucket_name):
     b2_tool.should_succeed(['get-bucket', bucket_name],)
     b2_tool.should_succeed(['ls', bucket_name],)
 
+    # Capabilities can be listed in any order. While this regex doesn't confirm that all three are present,
+    # in ensures that there are three in total.
+    failed_bucket_err = r'Deletion of file "test" \([^\)]+\) failed: unauthorized for ' \
+                        r'application key with capabilities ' \
+                        r"'(.*listFiles.*|.*listBuckets.*|.*readFiles.*){3}', " \
+                        r"restricted to bucket '%s' \(unauthorized\)" % bucket_name
+    b2_tool.should_fail(['rm', '--recursive', '--noProgress', bucket_name], failed_bucket_err)
+
     failed_bucket_err = r'ERROR: Application key is restricted to bucket: ' + bucket_name
     b2_tool.should_fail(['get-bucket', second_bucket_name], failed_bucket_err)
 
     failed_list_files_err = r'ERROR: Application key is restricted to bucket: ' + bucket_name
     b2_tool.should_fail(['ls', second_bucket_name], failed_list_files_err)
+
+    failed_list_files_err = r'ERROR: Application key is restricted to bucket: ' + bucket_name
+    b2_tool.should_fail(['rm', second_bucket_name], failed_list_files_err)
 
     # reauthorize with more capabilities for clean up
     b2_tool.should_succeed(
@@ -1428,6 +1451,77 @@ def test_sse_c(b2_tool, bucket_name):
     )
 
 
+@pytest.mark.skipif(
+    (sys.version_info.major, sys.version_info.minor) < (3, 8),
+    reason="License extraction doesn't work on older versions, and we're only "
+    "obliged to provide this "
+    "data in bundled and built packages."
+)
+@pytest.mark.parametrize('with_packages', [True, False])
+def test_license(b2_tool, with_packages):
+    license_text = b2_tool.should_succeed(
+        ['license'] + (['--with-packages'] if with_packages else [])
+    )
+
+    if with_packages:
+        full_license_re = re.compile(
+            r'Licenses of all modules used by b2, shipped with it in binary form:\r?\n'
+            r'\+-*\+-*\+\r?\n'
+            r'\|\s*Module name\s*\|\s*License text\s*\|\r?\n'
+            r'.*'
+            r'\+-*\+-*\+\r?\n', re.MULTILINE + re.DOTALL
+        )
+        full_license_text = next(full_license_re.finditer(license_text), None)
+        assert full_license_text
+        assert len(
+            full_license_text.group(0)
+        ) > 200_000  # we should know if the length of this block changes dramatically
+
+        license_summary_re = re.compile(
+            r'Summary of all modules used by b2, shipped with it in binary form:\r?\n'
+            r'\+-*\+-*\+-*\+-*\+-*\+\r?\n'
+            r'\|\s*Module name\s*\|\s*Version\s*\|\s*License\s*\|\s*Author\s*\|\s*URL\s*\|\r?\n'
+            r'.*'
+            r'\+-*\+-*\+-*\+-*\+-*\+\r?\n', re.MULTILINE + re.DOTALL
+        )
+        license_summary_text = next(license_summary_re.finditer(license_text), None)
+        assert license_summary_text
+        assert len(
+            license_summary_text.group(0)
+        ) > 15_000  # we should know if the length of this block changes dramatically
+
+    assert """b2 license:
+Backblaze wants developers and organization to copy and re-use our
+code examples, so we make the samples available by several different
+licenses.  One option is the MIT license (below).  Other options are
+available here:
+
+    https://www.backblaze.com/using_b2_code.html
+
+
+The MIT License (MIT)
+
+Copyright (c) 2015 Backblaze
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.""" in license_text.replace(os.linesep, '\n')
+
+
 def test_file_lock(b2_tool, application_key_id, application_key, b2_api):
     lock_disabled_bucket_name = b2_tool.generate_bucket_name()
     b2_tool.should_succeed(
@@ -2374,7 +2468,7 @@ def test_enable_file_lock_and_set_retention_at_once(b2_tool, b2_api, bucket_name
         [
             'update-bucket', bucket_name, '--defaultRetentionMode', 'compliance',
             '--defaultRetentionPeriod', '7 days'
-        ], 'ERROR: The bucket is not file lock enabled \(bucket_missing_file_lock\)'
+        ], r'ERROR: The bucket is not file lock enabled \(bucket_missing_file_lock\)'
     )
 
     # enable file lock and set retention at once